requirements.txt
Django>=4.0
python-dotenv
qdrant-client>=1.5.0
sentence-transformers
pdfminer.six